refactor(db): complete checkpoint coordinator facade
