+++ tools/readability-check/baseline.tsv
@@ -4,7 +4,6 @@ crates/laminar-core/src/cluster/control/leader_lease/mod.rs	6513	288	cluster-con
 crates/laminar-db/src/operator/interval_join/mod.rs	6012	689	operators	Interval-join planning, state mutation, and output hot loops remain co-located.	Extract cold validation and recovery without adding per-row cost; file may not grow.
 crates/laminar-db/src/operator/temporal_join/mod.rs	5965	496	operators	Temporal-join state transitions and record processing are performance-sensitive.	Extract cold planning and checkpoint paths after benchmark coverage; file may not grow.
 crates/laminar-db/src/db/mod.rs	5660	1258	database	Root database construction and cross-subsystem ownership still form a broad facade.	Move coherent registry and cluster-construction owners while preserving public paths; file may not grow.
-crates/laminar-db/src/checkpoint_coordinator/mod.rs	5597	424	checkpoint	Checkpoint admission, persistence, cleanup, and completion remain one protocol implementation.	Extract attempt lifecycle and cleanup owners under existing tests; file may not grow.
 crates/laminar-core/src/shuffle/transport/mod.rs	5212	354	shuffle	Transport authority, delivery accounting, retry, and connection tasks are coupled.	Separate cold connection/control ownership without changing frame hot paths; file may not grow.
 crates/laminar-db/src/operator_graph/mod.rs	5103	364	graph	Graph topology, execution, checkpoint state, and routing remain combined.	Extract topology and checkpoint-state owners after graph characterization; file may not grow.
 crates/laminar-db/src/rebalance/mod.rs	4371	413	rebalance	Assignment planning and fenced transition publication share ordering invariants.	Extract validation and immutable planning phases; file may not grow.
